gpu_limits: change min_{uniform,storage}_block_offset_alignment type to size_t

+++ libnopegl/src/gpu_limits.h
@@ -23,6 +23,7 @@
 #define GPU_LIMITS_H
 
 #include <stdint.h>
+#include <stdlib.h>
 
 struct gpu_limits {
     uint32_t max_texture_image_units;
@@ -32,9 +33,9 @@ struct gpu_limits {
     uint32_t max_compute_work_group_size[3];
     uint32_t max_compute_shared_memory_size;
     uint32_t max_uniform_block_size;
-    uint32_t min_uniform_block_offset_alignment;
+    size_t min_uniform_block_offset_alignment;
     uint32_t max_storage_block_size;
-    uint32_t min_storage_block_offset_alignment;
+    size_t min_storage_block_offset_alignment;
     uint32_t max_samples;
     uint32_t max_texture_dimension_1d;
     uint32_t max_texture_dimension_2d;
